+++ packages/client-wallets/src/sdk/balances.ts
@@ -229,9 +229,8 @@ export class Balances extends ClientSDK {
             if (nextCursor == null) {
                 return () => null;
             }
-
             const results = dlv(responseData, "cursor.data");
-            if (!results.length) {
+            if (!Array.isArray(results) || !results.length) {
                 return () => null;
             }